[#552] Auto fixed some linting errors on all source files.

+++ DisplayCAL/display_cal.py
@@ -7810,7 +7810,7 @@ class MainFrame(ReportFrame, BaseFrame, LUT3DMixin):
             0,
             hyperlink,
             flag=wx.ALIGN_LEFT | wx.ALIGN_CENTER_VERTICAL | wx.LEFT,
-            border=int(round(32 + 12)),
+            border=round(32 + 12),
         )
         dlg.description_txt_ctrl.SetFocus()
         dlg.sizer0.SetSizeHints(dlg)
